#101 Limit min width to original image width
The first implementation of the min width parameter did not care about whether the original image could actually support a crop of this size. The resulting image min width now is capped by the original image min width. If one needs to have a bigger resulting image it needs to be combined with a resize filter.